Make it separate
